refactor: migrate tsconfig alias resolution into separate file
