+++ vite.config.ts
@@ -1,8 +1,7 @@
-import { resolve } from 'node:path'
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import { visualizer } from 'rollup-plugin-visualizer'
-import tsconfig from './tsconfig.json'
+import { aliases_from_tsconfig } from './aliases'
 
 const defaultOptions = {
   host: 'local.dsde-dev.broadinstitute.org',
@@ -14,17 +13,6 @@ const defaultOptions = {
         cert: 'server.crt',
       },
   open: true,
-}
-
-function aliases_from_tsconfig() {
-  const paths: Record<string, string[]> = tsconfig.compilerOptions.paths
-  const aliases: Record<string, string> = {}
-  for (const [key, value] of Object.entries(paths)) {
-    const aliasKey = key.replace(/\/\*$/, '')
-    const aliasPath = resolve(__dirname, value[0].replace(/\/\*$/, ''))
-    aliases[aliasKey] = aliasPath
-  }
-  return aliases
 }
 
 // https://vitejs.dev/config/
